refactored fro lemmatizer
