+++ src/cltk/lemmatize/fro.py
@@ -1,8 +1,16 @@
-import re
-
-"""Rules are based on Brunot & Bruneau (1949).
+"""
+Lemmatizer for Old French.
+Rules are based on Brunot & Bruneau (1949).
 """
 
+import importlib.machinery
+import os
+import re
+
+from cltk.utils import CLTK_DATA_DIR
+
+__author__ = ["Natasha Voake <natashavoake@gmail.com>"]
+__license__ = "MIT License. See LICENSE."
 
 estre_replace = [
     (
@@ -25,9 +33,7 @@ avoir_replace = [
     )
 ]
 
-
 auxiliary_rules = estre_replace + avoir_replace
-
 
 first_conj_rules = [
     (
@@ -40,9 +46,7 @@ i_type_rules = [("i$|is$|it$|imes$|istes$|irent$|isse$", "ir")]
 
 u_type_rules = [("ui$|us$|ut$|umes$|ustes$|urent$|usse$", "oir")]
 
-
 verbal_rules = u_type_rules + i_type_rules + first_conj_rules
-
 
 regime_rules = [("on$|ain$", "e")]
 
@@ -50,12 +54,9 @@ plural_rules = [("ales$|aux$|aus$", "al"), ("s$", "")]
 
 masc_to_fem_rules = [("se$", "x"), ("ive$", "if"), ("ee$", "e")]
 
-
 french_nominal_rules = regime_rules + plural_rules + masc_to_fem_rules
 
-
 misc_rules = [("x$", "l"), ("z$", "t"), ("un$", "on"), ("eus$", "os"), ("^e$", "et")]
-
 
 determiner_rules = [
     ("^li$|^lo$|^le$|^la$|^les$", "le"),
@@ -65,7 +66,6 @@ determiner_rules = [
     ("^uns$|^une$|^unes$", "un"),
 ]
 
-
 reduction_rules = [
     ("d'", "de"),
     ("m'", "me"),
@@ -73,7 +73,6 @@ reduction_rules = [
     ("l'", "le"),
     ("qu'", "que"),
 ]
-
 
 patterns = (
     determiner_rules
@@ -85,7 +84,7 @@ patterns = (
 )
 
 
-def build_match_and_apply_functions(pattern, replace):
+def _build_match_and_apply_functions(pattern, replace):
     def matches_rule(word):
         return re.search(pattern, word)
 
@@ -95,12 +94,68 @@ def build_match_and_apply_functions(pattern, replace):
     return (matches_rule, apply_rule)
 
 
-rules = [
-    build_match_and_apply_functions(pattern, replace) for (pattern, replace) in patterns
-]
+rules = [_build_match_and_apply_functions(pattern, replace) for (pattern, replace) in patterns]
 
-
-def regex(token):
+def _apply_regex(token):
     for matches_rule, apply_rule in rules:
         if matches_rule(token):
             return apply_rule(token)
+
+
+class Lemmatizer(object):  # pylint: disable=too-few-public-methods
+
+    def __init__(self):
+
+        self.entries = self._load_entries()
+        self.forms_and_lemmas = self._load_forms_and_lemmas()
+        self.lemmas = [entry[0] for entry in self.entries]
+
+    def _load_entries(self):
+        """Check for availability of lemmatizer for French."""
+
+        rel_path = os.path.join(
+            CLTK_DATA_DIR, "fro", "text", "fro_data_cltk", "entries.py"
+        )
+        path = os.path.expanduser(rel_path)
+        loader = importlib.machinery.SourceFileLoader("entries", path)
+        module = loader.load_module()
+        entries = module.entries
+
+        return entries
+
+    def _load_forms_and_lemmas(self):
+
+        rel_path = os.path.join(
+            CLTK_DATA_DIR, "fro", "text", "fro_data_cltk", "forms_and_lemmas.py"
+        )
+        path = os.path.expanduser(rel_path)
+        loader = importlib.machinery.SourceFileLoader("forms_and_lemmas", path)
+        module = loader.load_module()
+        forms_and_lemmas = module.forms_and_lemmas
+
+        return forms_and_lemmas
+
+    def lemmatize(self, token: str) -> str:
+        """
+        Lemmatize French words by replacing input words with corresponding
+        values from a replacement list.
+        >>> lemmatizer = Lemmatizer()
+        >>> lemmatizer.lemmatize('out')
+        ['avoir']
+        """
+        #check for a match between token and list of lemmas
+        if token in self.lemmas:
+            return [token]
+
+        #if no match check for a match between token and list of lemma forms
+        lemmas = [lemma for lemma, forms in self.forms_and_lemmas.items() if token in forms]
+        if lemmas:
+            return lemmas
+
+        #if no match apply regular expressions and check for a match against the list of lemmas again
+        mod_token = _apply_regex(token)
+        if mod_token in self.lemmas:
+            return [mod_token]
+
+        return [token]
+
